Complete REST API with filtering, Pagenation, Signals, Celery, Mails and Soft Delete

+++ apis/viewsets.py
@@ -8,6 +8,29 @@ from django.contrib.auth import get_user_model
 from hotels.models import Hotel, Room, Classification, Reservation
 import django_filters.rest_framework
 from django_filters.rest_framework import DjangoFilterBackend
+from datetime import datetime, timedelta
+from django.db.models.signals import post_delete
+from django.template.loader import get_template
+from django.core.mail import EmailMultiAlternatives
+
+
+def send_notification(sender, instance, **kwargs):
+
+    subject = 'Reservation Expired'
+    from_mail = 'aa6653312@gmail.com'
+    to_mail = [instance.user.email,]
+    context = {
+        'name': instance.user.username,
+        'expire_date': instance.expired_at,
+    }
+    message = get_template('mesages/message-template.html').render(context)
+
+    msg = EmailMultiAlternatives(subject, message, from_mail, [to_mail])
+    msg.attach_alternative(message, "text/html")
+    msg.send()
+
+
+post_delete.connect(send_notification, sender=Reservation)
 
 
 
@@ -25,7 +48,7 @@ class UserViewSet(viewsets.ModelViewSet):
 class HotelViewSet(viewsets.ModelViewSet):
     queryset = Hotel.objects.all()
     serializer_class = HotelSerializer
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
+    permission_classes = (permissions.IsAdminUser,)
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ['name', 'stars']
     search_fields = ['^name', '=stars']
@@ -34,7 +57,7 @@ class HotelViewSet(viewsets.ModelViewSet):
 class DeletedHotelViewSet(viewsets.ModelViewSet):
     queryset = Hotel.objects.deleted_only()
     serializer_class = HotelSerializer
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
+    permission_classes = (permissions.IsAdminUser,)
 
 
 #######################################################
@@ -43,7 +66,7 @@ class DeletedHotelViewSet(viewsets.ModelViewSet):
 class RoomViewSet(viewsets.ModelViewSet):
     queryset = Room.objects.all()
     serializer_class = RoomSerializer
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
+    permission_classes = (permissions.IsAdminUser,)
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ['number', 'hotel', 'is_avaliable', 'price', 'classification']
     search_fields = ['=number', '=hotel', '=is_avaliable', 'price', '=classification']
@@ -52,7 +75,7 @@ class RoomViewSet(viewsets.ModelViewSet):
 class AvailableRoomViewSet(viewsets.ModelViewSet):
     queryset = Room.objects.filter(is_avaliable=True).all()
     serializer_class = RoomSerializer
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
+    permission_classes = (permissions.AllowAny,)
     filter_backends = [DjangoFilterBackend]
     filterset_fields = ['number', 'hotel', 'price', 'classification']
     search_fields = ['=number', '=hotel', 'price', '=classification']
@@ -61,7 +84,7 @@ class AvailableRoomViewSet(viewsets.ModelViewSet):
 class DeletedRoomViewSet(viewsets.ModelViewSet):
     queryset = Room.objects.deleted_only()
     serializer_class = RoomSerializer
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
+    permission_classes = (permissions.IsAdminUser,)
 
 
 
@@ -94,17 +117,24 @@ class ReservationViewSet(viewsets.ModelViewSet):
     search_fields = ['=room', '=user', 'created_at', 'started_at', 'expired_at']
     # permission_classes = (permissions.IsAuthenticated,)
 
-    """ making room availablr for future reservations"""
+    """ ensuring that the deletion of the reservation performs before at
+    least 3 days of reservationmaking room availablr for future reservations"""
     def destroy(self, request, *args, **kwargs):
         obj = self.get_object()
-        room = obj.room
-        room.is_avaliable = True
-        room.save()
-        obj.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if obj.started_at.month >= datetime.now().month:
+            if obj.started_at.day > datetime.now().day + 3:
+                room = obj.room
+                room.is_avaliable = True
+                room.save()
+                obj.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response(status=status.HTTP_406_NOT_ACCEPTABLE)
+
+
 
 
 class DeletedReservationViewSet(viewsets.ModelViewSet):
     queryset = Reservation.objects.deleted_only()
     serializer_class = ReservationSerializer
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
+    # permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
